+++ diskord/__init__.py
@@ -13,7 +13,7 @@ __title__ = "discord"
 __author__ = "Rapptz"
 __license__ = "MIT"
 __copyright__ = "Copyright 2015-present Rapptz"
-__version__ = "2.6.0a"
+__version__ = "2.6.1"
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)
 
@@ -72,7 +72,7 @@ class VersionInfo(NamedTuple):
 
 
 version_info: VersionInfo = VersionInfo(
-    major=2, minor=5, micro=0, releaselevel="alpha", serial=0
+    major=2, minor=6, micro=1, releaselevel="final", serial=0
 )
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())