+++ diskord/guild.py
@@ -3245,52 +3245,6 @@ class Guild(Hashable):
 
     # Application commands management
 
-    def get_application_command(self, command_id: int, /) -> ApplicationCommand:
-        """
-        Gets an application command for this guild with the command's ID.
-
-        The command for provided ID must belong to this guild.
-
-        Parameters
-        ----------
-
-        command_id: :class:`int`
-            The ID of the command to get.
-
-        Returns
-        -------
-
-        :class:`ApplicationCommand`
-            The required command.
-        """
-        for command in self._application_commands:
-            if command.id == command_id and self.id == command.guild_id:
-                return command
-
-    def remove_application_command(self, command_id: int, /):
-        """
-        Removes an application command from internal application commands list for this
-        guild.
-
-        The command for provided ID must belong to this guild.
-
-        If any error is raised during command removal, It is silently suppressed.
-
-        Parameters
-        -----------
-
-        command_id: :class:`int`
-            The ID of the command to remove.
-        """
-        from .application_commands import ApplicationCommand  # due to circular import
-
-        for command in self._state._application_commands:
-            if command.id == command_id and self.id == command.guild_id:
-                try:
-                    return self._state._application_commands.remove(command)
-                except ValueError:
-                    return
-
     async def fetch_application_command(self, command_id: int, /) -> ApplicationCommand:
         """|coro|
 
@@ -3322,15 +3276,12 @@ class Guild(Hashable):
             application_id=self._state.user.id,
             guild_id=self.id,
         )
+        return ApplicationCommand(data)
 
     async def fetch_application_command_permissions(self):
         """|coro|
 
-        Fetches the permissions of an application command in this guild.
-
-        .. note::
-            This function is an API call, Use :func:`get_application_command_permissions`
-            for general usage.
+        Fetches the permissions of all application commands in this guild.
 
         Returns
         -------
@@ -3338,24 +3289,16 @@ class Guild(Hashable):
         List[:class:`ApplicationCommandGuildPermissions`]
             The list of permissions for each command.
         """
-        response = await self.http.get_guild_application_command_permissions(
+        data = await self.http.get_guild_application_command_permissions(
             application_id=self.user.id,
             guild_id=self.id,
         )
-        for perm in response:
-            perm["permissions"] = [
-                ApplicationCommandPermission(**p) for p in perm["permissions"]
-            ]
-        return ApplicationCommandGuildPermissions(**response)
+        return [ApplicationCommandGuildPermissions(perm) for perm in data]
 
     async def fetch_application_command_permission(self, command_id: int, /):
         """|coro|
 
         Fetches the permissions of an application command in this guild by command ID.
-
-        .. note::
-            This function is an API call, Use :func:`get_application_command_permission`
-            for general usage.
 
         Returns
         -------
@@ -3363,12 +3306,9 @@ class Guild(Hashable):
         :class:`ApplicationCommandGuildPermissions`
             The permissions for the provided command.
         """
-        response = await self.http.get_application_command_permissions(
+        data = await self.http.get_application_command_permissions(
             application_id=self.user.id,
             guild_id=self.id,
             command_id=command_id,
         )
-        response["permissions"] = [
-            ApplicationCommandPermission(**perm) for perm in response["permissions"]
-        ]
-        return ApplicationCommandGuildPermissions(**response)
+        return ApplicationCommandGuildPermissions(data)